Add functionality to fetch and display drinks data
This update includes the creation of a new DrinksService class that interacts with an API to fetch various drinks categories and drinks data. A UserInput class has been added to get user input, and a TableVisualisationEngine class to display data in a tabular format. In addition, new validation for user input and necessary models to handle data have been incorporated. Package dependencies have also been updated.